prevent inlining of fib calls
The measurements were distorted by different degrees of inlining
of recursive fib calls. We now explicitly forbid that inlinine.
In a way that penalizes the approaches, but otherwise comparisons
are misleading. Conceptually, all affected approaches should allow
for the same degree of inlining (but in practice the compiler
choses differently).